feat(user-endpoints): refactor implementation
- define handleErrors function to absctract error handling
- refactor button component to have spinner
- return only token in graphQL requests
