+++ src/components/index.js
@@ -1,0 +1,9 @@
+import button from './button';
+import formGroup from './formGroup';
+import socialAuths from './socialAuths';
+import modal from './modal';
+
+export const Button = button;
+export const FormGroup = formGroup;
+export const SocialAuths = socialAuths;
+export const Modal = modal;
