+++ src/graphql/mutations.js
@@ -16,11 +16,6 @@ export const createLink = gql`
 export const signup = gql`
   mutation Signup($name: String!, $email: String!, $password: String!) {
     signup(name: $name, email: $email, password: $password) {
-      user {
-        id
-        name
-        email
-      }
       token
     }
   }
@@ -29,11 +24,6 @@ export const signup = gql`
 export const login = gql`
   mutation Login($email: String!, $password: String!) {
     signin(email: $email, password: $password) {
-      user {
-        id
-        name
-        email
-      }
       token
     }
   }
